springsecurity

+++ board/src/main/java/com/joeun/board/controller/FileController.java
@@ -15,6 +15,7 @@ import org.springframework.util.FileCopyUtils;
 import org.springframework.web.bind.annotation.DeleteMapping;
 import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.PathVariable;
+import org.springframework.web.bind.annotation.RequestBody;
 import org.springframework.web.bind.annotation.RequestMapping;
 
 import com.joeun.board.dto.Files;
@@ -76,11 +77,19 @@ public class FileController {
 
 
     @DeleteMapping("")
+    // public ResponseEntity<String> deleteFile(@RequestBody Files file) throws Exception {
     public ResponseEntity<String> deleteFile(Files file) throws Exception {
+        log.info("[DELETE] - /file");
         int fileNo = file.getFileNo();
+        log.info("fileNo : " + fileNo);
+        if( fileNo == 0 )
+            return new ResponseEntity<String>("FAIL", HttpStatus.BAD_REQUEST);       
 
         int result = fileService.delete(fileNo);
 
+        if( result == 0 )
+            return new ResponseEntity<String>("FAIL", HttpStatus.OK);    
+        
         return new ResponseEntity<String>("SUCCESS", HttpStatus.OK);
     }
